break out review step
